Migrate all remaining alerters out of alerts.py
